refactor(config): simplify configuration file handling
Configuration file creation, etc., is only necessary when actually
saving changed settings.  This will be driven by the user interface.

If the configuration file doesn't exist, silently skip it.  Only throw
an error at load time if the file exists but cannot be read for
whatever reason, on the assumption that if the configuration file
exists, it is supposed to be readable.