fix(database): de-serialize a system's factions first
Sometimes, a station's data on a faction can be invalid or missing.
For example, Pettitt Metalurgic Base on 36 Ophiuchi A 1 a in the test
galaxy data claims the 36 Ophiuchi Autocracy faction is
Federation-aligned due to these journal entries:

https://edgalaxydata.space/eddn-lookup/extract.php?filename=Journal.Docked-2022-09-19.jsonl.bz2&lineno=7315

https://edgalaxydata.space/eddn-lookup/extract.php?filename=Journal.Docked-2023-01-01.jsonl.bz2&lineno=11071

While the Marshmallow documentation for the knob mentions
serialization, setting `ordered = True` and listing the schema fields
in this order appears to lead to the desired result---that the
system's factions get created (and memoized) from the system's
`factions` field first.  Then when Marshmallow processes the system's
stations and bodies, any faction references will use the cached
Faction object.

https://marshmallow.readthedocs.io/en/stable/marshmallow.schema.html#marshmallow.schema.Schema.Meta